Refactor: reorganize component file structure to follow module-based architecture
