--- frontend/src/components/UI/CodeEditor/CodeEditor.module.scss
@@ -1,6 +1,0 @@
-@use '../../../styles/variables' as *;
-@use '../../../styles/mixins' as *;
-
-.codeEditor {
-  /* Your styles here */
-}
