--- frontend/src/components/UI/index.js
@@ -1,7 +1,0 @@
-// UI Components Barrel Exports
-export { default as Breadcrumb } from './Breadcrumb/Breadcrumb';
-export { default as CodeEditor } from './CodeEditor/CodeEditor';
-export { default as LanguageSetting } from './LanguageSetting/LanguageSetting';
-export { default as NavigationBar } from './NavigationBar/NavigationBar';
-export { default as SettingPanel } from './SettingPanel/SettingPanel';
-export { default as UserStatus } from './UserStatus/UserStatus';
