--- frontend/src/components/index.js
@@ -1,8 +1,0 @@
-// Main Components Barrel Exports
-
-export * from './UI';
-
-export * from './Header';
-
-// Footer component
-export { default as Footer } from './Footer/Footer';
