+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App.jsx'
-import './styles/main.scss';
+import './shared/styles/main.scss';
 import './i18n';
 
 // 會找到 index.html 中 #root 的 HTML 元素，並將App渲染到DOM 元素中
