+++ frontend/src/modules/core/components/CodeEditor/CodeEditor.jsx
