+++ frontend/src/modules/core/components/CodeEditor/CodeEditor.module.scss
@@ -1,0 +1,6 @@
+@use '../../../../shared/styles/variables' as *;
+@use '../../../../shared/styles/mixins' as *;
+
+.codeEditor {
+  /* Your styles here */
+}
