+++ frontend/src/modules/platform/components/LanguageSetting/LanguageSetting.module.scss
