+++ frontend/src/modules/user/components/SettingPanel/SettingPanel.module.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/variables' as *;
-@use '../../../styles/mixins' as *;
+@use '../../../../shared/styles/variables' as *;
+@use '../../../../shared/styles/mixins' as *;
 
 .backdrop {
   position: fixed;
