+++ frontend/src/pages/About/About.module.scss
@@ -1,6 +1,3 @@
-@use '../../styles/variables' as *;
-@use '../../styles/mixins' as *;
-
 .about {
   min-height: 100vh;
   background: var(--bg-primary);
