+++ frontend/src/pages/Explorer/Explorer.module.scss
@@ -1,6 +1,3 @@
-@use '../../styles/variables' as *;
-@use '../../styles/mixins' as *;
-
 .explorer {
   min-height: 100vh;
   background: var(--bg-primary);
