+++ frontend/src/pages/LearningDashboard/LearningDashboard.module.scss
@@ -1,6 +1,3 @@
-@use '../../styles/variables' as *;
-@use '../../styles/mixins' as *;
-
 .tempTutorialSection {
   margin-top: $spacing-xl;
