+++ frontend/src/pages/Tutorial/Tutorial.module.scss
@@ -1,6 +1,3 @@
-@use '../../styles/variables' as *;
-@use '../../styles/mixins' as *;
-
 .tutorialPage {
   min-height: calc(100vh - 64px);
   background: $dark-bg-primary;
