+++ frontend/src/pages/Tutorial/components/AlgorithmInfo/AlgorithmInfo.module.scss
@@ -1,6 +1,3 @@
-@use '../../../../styles/variables' as *;
-@use '../../../../styles/mixins' as *;
-
 .algorithmInfo {
   /* Your styles here */
 }
