+++ frontend/src/pages/Tutorial/components/AnimationArea/AnimationArea.module.scss
@@ -1,6 +1,3 @@
-@use '../../../../styles/variables' as *;
-@use '../../../../styles/mixins' as *;
-
 .animationArea {
   /* Your styles here */
 }
