+++ frontend/src/pages/Tutorial/components/ControlBar/ControlBar.module.scss
@@ -1,6 +1,3 @@
-@use '../../../../styles/variables' as *;
-@use '../../../../styles/mixins' as *;
-
 .controlBar {
   /* Your styles here */
 }
