+++ frontend/src/pages/Tutorial/components/VariableStatus/VariableStatus.module.scss
@@ -1,6 +1,3 @@
-@use '../../../../styles/variables' as *;
-@use '../../../../styles/mixins' as *;
-
 .variableStatus {
   /* Your styles here */
 }
