+++ frontend/src/shared/components/Footer/Footer.jsx
