+++ frontend/src/shared/components/Footer/Footer.module.scss
