+++ frontend/src/shared/components/Header/AuthHeader.jsx
