+++ frontend/src/shared/components/Header/MainHeader.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styles from './Header.module.scss';
-import NavigationBar from '../UI/NavigationBar/NavigationBar';
-import LanguageSetting from '../UI/LanguageSetting/LanguageSetting';
-import UserStatus from '../UI/UserStatus/UserStatus';
+import NavigationBar from '../NavigationBar/NavigationBar';
+import LanguageSetting from '../../../modules/platform/components/LanguageSetting/LanguageSetting';
+import UserStatus from '../../../modules/user/components/UserStatus/UserStatus';
 import logo from '/images/codePulse_logo.png';
 
 function MainHeader() {
