+++ frontend/src/shared/components/NavigationBar/NavigationBar.jsx
