+++ frontend/src/shared/components/NavigationBar/NavigationBar.module.scss
