+++ frontend/src/shared/components/index.js
@@ -1,0 +1,6 @@
+// Shared Components Barrel Exports
+export { default as Breadcrumb } from './Breadcrumb/Breadcrumb';
+export { default as NavigationBar } from './NavigationBar/NavigationBar';
+export { default as MainHeader } from './Header/MainHeader';
+export { default as AuthHeader } from './Header/AuthHeader';
+export { default as Footer } from './Footer/Footer';
