+++ frontend/src/shared/layouts/AuthLayout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import { AuthHeader } from '@/components';
+import { AuthHeader } from '../components';
 
 function AuthLayout() {
   return (
