+++ frontend/src/shared/styles/_common.scss
