+++ frontend/src/shared/styles/_variables.scss
