+++ frontend/src/shared/styles/main.scss
